Add restart-bluetooth
Usage:
    restart-bluetooth.sh [wifi]

With the "wifi" argument, the Wi-Fi device is also turned off 
then on again after a short delay.

With MacOS X 10.13, a lot of the time my headphones won't be seen 
/ connected.  This script solves that problem for me much more
reliably than anything else I've found.

Having both Wi-Fi and Bluetooth off for a short time seems to 
clear up most of the issues I'm seeing. But that's very
disruptive so I tend to try just resetting the Bluetooth first.
Sometimes, I need to do both to get it working again.

The fact that my MacBook Pro doesn't initiate a connection to my 
headphones after enabling Bluetooth is extremely annoying!!

Without this script I'm forced to hit the buttons on the 
headphones myself to try to connect multiple times since the 
connection often fails in this direction and if I don't the
headphones power off without any sounds leaving me to guess if
they are on or not while I'm trying to reconnect.